+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [1.5.0]
+### Removed
+- 'reference' from the payment actions
+### Added
+- Unique identifier to the payment actions
+
 ## [1.4.0]
 ### Added
 - Implemented 'reference' to the payment actions